+++ BonApp/src/main/resources/templates/erreur.html
@@ -1,0 +1,13 @@
+<!DOCTYPE html>
+<html lang="fr" xmlns:th="http://www.thymeleaf.org">
+<head>
+    <meta charset="UTF-8">
+    <title>Error</title>
+</head>
+<body>
+    <h1>Oups... Il y a eu un probleme</h1>
+
+    <a href="/" style="">Accueil</a>
+
+</body>
+</html>
